Выносит цикл с генерацией фото в отдельную функцию
